requirements.txt
uvicorn~=0.30.1
python-dotenv~=1.0.1
aiogram~=3.10.0
fastapi~=0.111.0
tortoise-orm~=0.21.4
Pillow~=10.5.0